Move ElementPackingUtils to Encoding
