+++ compiler/src/iree/compiler/Dialect/Encoding/Utils/BUILD.bazel
@@ -15,14 +15,18 @@ package(
 iree_compiler_cc_library(
     name = "Utils",
     srcs = [
+        "ElementPackingUtils.cpp",
         "Utils.cpp",
     ],
     hdrs = [
+        "ElementPackingUtils.h",
         "Utils.h",
     ],
     deps = [
         "//compiler/src/iree/compiler/Dialect/Encoding/IR",
+        "//compiler/src/iree/compiler/Dialect/Util/IR",
         "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:ArithDialect",
         "@llvm-project//mlir:IR",
         "@llvm-project//mlir:InferTypeOpInterface",
         "@llvm-project//mlir:LinalgDialect",
