+++ compiler/src/iree/compiler/Dialect/Stream/IR/BUILD.bazel
@@ -68,7 +68,6 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Dialect/Encoding/Utils",
         "//compiler/src/iree/compiler/Dialect/Util/IR",
         "//compiler/src/iree/compiler/Utils",
-        "//compiler/src/iree/compiler/Utils:PackingUtils",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:ArithDialect",
         "@llvm-project//mlir:ArithUtils",
