+++ compiler/src/iree/compiler/Dialect/Stream/IR/CMakeLists.txt
@@ -58,7 +58,6 @@ iree_cc_library(
     iree::compiler::Dialect::Encoding::Utils
     iree::compiler::Dialect::Util::IR
     iree::compiler::Utils
-    iree::compiler::Utils::PackingUtils
   PUBLIC
 )
