+++ compiler/src/iree/compiler/Dialect/Stream/Transforms/BUILD.bazel
@@ -80,7 +80,6 @@ iree_compiler_cc_library(
         "//compiler/src/iree/compiler/Dialect/Util/IR",
         "//compiler/src/iree/compiler/Dialect/Util/Transforms",
         "//compiler/src/iree/compiler/Utils",
-        "//compiler/src/iree/compiler/Utils:PackingUtils",
         "@llvm-project//llvm:Support",
         "@llvm-project//mlir:AffineDialect",
         "@llvm-project//mlir:Analysis",
