+++ compiler/src/iree/compiler/Dialect/Stream/Transforms/CMakeLists.txt
@@ -91,7 +91,6 @@ iree_cc_library(
     iree::compiler::Dialect::Util::IR
     iree::compiler::Dialect::Util::Transforms
     iree::compiler::Utils
-    iree::compiler::Utils::PackingUtils
   PUBLIC
 )
